requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Mpesa==1.2.8
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.24
typing_extensions==4.9.0
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
